Adding Signal_handler example
This program demonstrates the use of signal handlers in UNIX. Each signal is represented as a single bit flag, and thus signals of the same type cannot be
queued. When a process is scheduled for execution, it checks these signal flags and performs
corresponding actions for posted signals before returning to user level. The signals can be
handled by predefined (default) actions (e.g., termination for SIGTERM), can be caught by
user-defined signal handlers, or (in some cases) can be ignored by the process. User-defined
signal handler functions execute at user level and in the context of the target process.